video owner commit

+++ src/controller/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 // export const home = (req, res) => {
 // 	Video.finde({}, (error, videos) => {
@@ -13,7 +14,8 @@ export const home = async(req, res) => {
 
 export const watch = async(req, res) => {
 	const { id } = req.params;
-	const video = await Video.findById(id);
+	const video = await Video.findById(id).populate("owner");
+	
 	if (!video) {
 		return res.status(404).render("404", { pageTitle : "Video not found" });
 	}
@@ -21,19 +23,27 @@ export const watch = async(req, res) => {
 }
 export const getEdit = async (req, res) => {
 	const { id } = req.params;
+	const {user: {_id}} = req.session;
 	const video = await Video.findById(id);
 	if (!video) {
 		return res.status(404).render("404", { pageTitle : "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
 	}
 	res.render("edit", {pageTitle : `Edit ${video.title}`, video});
 }
 
 export const postEdit = async (req, res) => {
+	const {user: {_id}} = req.session;
 	const { id } = req.params;
 	const { title, description, hashtags } = req.body;
 	const video = await Video.exists({ _id: id });
 	if (!video) {
 		return res.render("404", { pageTitle : "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
 	}
 	await Video.findByIdAndUpdate(id, {
 		title, 
@@ -53,15 +63,22 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
+	const {
+		user:{ _id }
+	} = req.session;
 	const { path:fileUrl } = req.file;
 	const { title, description, hashtags } = req.body;
 	try{
-		await Video.create({
+		const newVideo = await Video.create({
 			title,
 			description,
 			fileUrl,
+			owner: _id,
 			hashtags: Video.formatHashtags(hashtags),
 		});
+		const user = await User.findById(_id);
+		user.videos.push(newVideo._id);
+		user.save();
 		return res.redirect("/");
 	} catch(error) {
 		console.log(error);
@@ -73,7 +90,15 @@ export const postUpload = async (req, res) => {
 }
 
 export const deleteVideo = async (req, res) => {
+	const {user: {_id}} = req.session;
 	const { id } = req.params;
+	const video = await Video.findById(id);
+	if (!video) {
+		return res.status(404).render("404", { pageTitle : "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+	}
 	await Video.findByIdAndDelete(id);
 	return res.redirect("/");
 }
